do not attempt to add field if field already on canvas
