+++ Stitch/Graph/Node/View/NodesView.swift
@@ -71,9 +71,6 @@ struct CanvasEdgesViewModifier: ViewModifier {
                 CandidateEdgesView(graph: graph)
             }
             .overlay {
-//                EdgeDrawingView(graph: graph,
-//                                edgeDrawingObserver: graph.edgeDrawingObserver)
-                
                 if let edgeEditingState = graph.edgeEditingState {
                     EdgeInputLabelsView(document: document,
                                         graph: graph,
