+++ Stitch/Graph/View/GraphBaseView.swift
@@ -126,9 +126,8 @@ struct GraphBaseView: View {
                 }
         } // ZStack
         
-        .modifier(ActivelyDrawnEdgeThatCanEnterInspector(
-            graph: graph,
-            scale: document.graphMovement.zoomData))
+        .modifier(ActivelyDrawnEdge(graph: graph,
+                                    scale: document.graphMovement.zoomData))
         
         .coordinateSpace(name: Self.coordinateNamespace)
         .background {
@@ -146,209 +145,6 @@ struct GraphBaseView: View {
             } // GeometryReader
         } // .background
     }
-}
-
-struct ActivelyDrawnEdgeThatCanEnterInspector: ViewModifier {
-    
-    @Bindable var graph: GraphState
-    let scale: CGFloat
-        
-    // finds canvas input OR inspector input/field
-    @MainActor
-    func findEligibleInput(_ drawingObserver: EdgeDrawingObserver,
-                                         draggedOutputRect: CGRect,
-                                         geometry: GeometryProxy,
-                                         preferences: [EdgeDraggedToInspector: Anchor<CGRect>]) -> EmptyView {
-        
-        guard let drawingGesture = drawingObserver.drawingGesture else {
-            return EmptyView()
-        }
-        
-        let dragLocation = drawingGesture.cursorLocationInGlobalCoordinateSpace
-        
-        var nearestInspectorInputs = [LayerInputType]()
-        
-        for preference in preferences {
-            if case let .inspectorInputOrField(layerInputType) = preference.key,
-               // Note: `areNear` *already* expands the 'hit area'
-//                areNear(geometry[preference.value].origin,
-                areNear(geometry[preference.value].mid,
-                        dragLocation,
-                        isInspectorInputOrFieldDetection: true) {
-                
-                log("findEligibleInspectorFieldOrRow: WAS NEAR: layerInputType: \(layerInputType)")
-                nearestInspectorInputs.append(layerInputType)
-            }
-        } // for preference in ...
-                
-        
-        DispatchQueue.main.async { [weak graph] in
-            
-            guard let graph = graph,
-                  graph.edgeDrawingObserver.drawingGesture.isDefined else {
-                log("findEligibleInspectorFieldOrRow: no longer have an output drag")
-                return
-            }
-            
-            if let outputNodeId = drawingGesture.outputId.graphItemType.getCanvasItemId,
-               let dragLocationInNodesViewCoordinateSpace = graph.dragLocationInNodesViewCoordinateSpace {
-                graph.findEligibleCanvasInput(
-                    cursorLocation: dragLocationInNodesViewCoordinateSpace,
-                    cursorNodeId: outputNodeId)
-            }
-            
-            let hadEligibleInspectorInputOrField = drawingObserver.nearestEligibleEdgeDestination?.getInspectorInputOrField.isDefined ?? false
-            
-            if nearestInspectorInputs.isEmpty,
-               hadEligibleInspectorInputOrField {
-                log("findEligibleInspectorFieldOrRow: NO inspector inputs/fields")
-                drawingObserver.nearestEligibleEdgeDestination = nil
-            } else if let nearestInspectorInput = nearestInspectorInputs.last {
-                log("findEligibleInspectorFieldOrRow: found inspector input/field: \(nearestInspectorInput)")
-                drawingObserver.nearestEligibleEdgeDestination = .inspectorInputOrField(nearestInspectorInput)
-            }
-            
-            // After we've set or wiped the nearestEligible input,
-            // *animate* the port color change:
-            withAnimation(.linear(duration: DrawnEdge.ANIMATION_DURATION)) {
-                graph
-                    .getOutputRowObserver(drawingGesture.outputId.asNodeIOCoordinate)?
-                    .updateRowViewModelsPortColor(selectedEdges: graph.selectedEdges,
-                                                  selectedCanvasItems: graph.selectedCanvasItems,
-                                                  drawingObserver: drawingObserver)
-            }
-        }
-                
-        return EmptyView()
-    }
-    
-    @Environment(\.appTheme) var theme
-        
-    // Note: the rules for the color of an actively dragged edge are simple:
-    // gray if no eligible input, else highlighted-loop if a loop, else highlighted.
-    @MainActor
-    func color(_ outputRowViewModel: OutputNodeRowViewModel) -> PortColor {
-        if !graph.edgeDrawingObserver.nearestEligibleEdgeDestination.isDefined {
-            return .noEdge
-        } else if (outputRowViewModel.rowDelegate?.hasLoopedValues ?? false) {
-            return .highlightedLoopEdge
-        } else {
-            return .highlightedEdge
-        }
-    }
-    
-    // TODO: is this really appropriate even for actively dragged circuit edges? ... there's no such thing as an an
-    @MainActor
-    var inputAnchorData: EdgeAnchorDownstreamData? {
-        
-        // just need the nearest eligible canvas/inspector's input node row view model
-        
-        switch graph.edgeDrawingObserver.nearestEligibleEdgeDestination {
-            
-        case .none:
-            return nil
-        
-        case .canvasInput(let inputNodeRowViewModel):
-            return EdgeAnchorDownstreamData(from: inputNodeRowViewModel)
-        
-        case .inspectorInputOrField(let layerInputType):
-            if let firstFocusedLayer = graph.inspectorFocusedLayers.first,
-               let nearestEligibleInputOrFieldRowViewModel: InputNodeRowViewModel = graph.getInputRowViewModel(for:
-                     .init(graphItemType: .layerInspector(.keyPath(layerInputType)),
-                           nodeId: firstFocusedLayer.asLayerNodeId.asNodeId,
-                           // All layer inputs/input-fields have 0
-                           portId: 0)) {
-                return EdgeAnchorDownstreamData(from: nearestEligibleInputOrFieldRowViewModel)
-            } else {
-                return nil
-            }
-        }
-    }
-    
-    @MainActor @ViewBuilder
-    func body(content: Content) -> some View {
-        // TODO: is this acceptable for perf?
-        content
-            .overlayPreferenceValue(EdgeDraggedToInspectorPreferenceKey.self) { preferences in
-                GeometryReader { geometry in
-                    if let drawingGesture = graph.edgeDrawingObserver.drawingGesture,
-                       // The output from which the currently-dragged edge originates
-                       let draggedOutputPref = preferences[.draggedOutput(drawingGesture.outputId.asNodeIOCoordinate)],
-                       let outputRowViewModel = self.graph.getOutputRowViewModel(for: drawingGesture.outputId) {
-                        
-                        // Location of dragged edge's end, i.e. user's cursor position
-                        let draggedOutputRect: CGRect = geometry[draggedOutputPref]
-                        
-                        // Always draw the
-                        let pointTo = drawingGesture.cursorLocationInGlobalCoordinateSpace
-                        
-                        
-                        if let downstreamNode = graph.getNode(drawingGesture.outputId.nodeId),
-                           let upstreamCanvasItem = outputRowViewModel.canvasItemDelegate,
-                            let outputAnchorData = EdgeAnchorUpstreamData(
-                                from: upstreamCanvasItem.outputPortUIViewModels,
-                                upstreamNodeId: upstreamCanvasItem.id.nodeId,
-                                inputRowViewModelsOnDownstreamNode: downstreamNode.allInputViewModels),
-                           let outputPortAddress = outputRowViewModel.portUIViewModel.portAddress,
-                           let outputNodeId = outputRowViewModel.canvasItemDelegate?.id
-//                            ,
-//                           let pointFrom = outputRowViewModel.portUIViewModel.anchorPoint
-                        {
-                            
-                            // Note:
-                            let pointFrom = draggedOutputRect.mid
-                            
-                            logInView("EdgeFromDraggedOutputView: pointFrom: \(pointFrom)")
-                            logInView("EdgeFromDraggedOutputView: pointTo: \(pointTo)")
-                            
-                            let edge = PortEdgeUI(from: outputPortAddress,
-                                                  to: .init(portId: -1, // Nonsense
-                                                            canvasId: outputNodeId))
-                            
-                            let color = self.color(outputRowViewModel)
-                            
-                            EdgeView(edge: edge,
-                                     pointFrom: pointFrom,
-                                     pointTo: pointTo,
-                                     color: color.color(theme),
-                                     isActivelyDragged: true, // always true for actively-dragged edge
-                                     firstFrom: outputAnchorData.firstUpstreamOutput.anchorPoint ?? .zero,
-                                     firstTo: inputAnchorData?.firstInput.anchorPoint ?? .zero,
-                                     lastFrom: outputAnchorData.lastUpstreamRowOutput.anchorPoint ?? .zero,
-                                     lastTo: inputAnchorData?.lastInput.anchorPoint ?? .zero,
-                                     firstFromWithEdge: outputAnchorData.firstConnectedUpstreamOutput?.anchorPoint?.y,
-                                     lastFromWithEdge: outputAnchorData.lastConnectedUpstreamOutput?.anchorPoint?.y,
-                                     firstToWithEdge: inputAnchorData?.firstConnectedInput.anchorPoint?.y,
-                                     lastToWithEdge: inputAnchorData?.lastConectedInput.anchorPoint?.y,
-                                     totalOutputs: outputAnchorData.totalOutputs,
-                                     // we never animate the actively dragged edge
-                                     edgeAnimationEnabled: false,
-                                     edgeScaleEffect: scale)
-                            .animation(.linear(duration: DrawnEdge.ANIMATION_DURATION),
-                                       value: color)
-                        }
-                        
-//                        // Render the actively-drawn-edge
-//                        CurveLine(from: draggedOutputRect.mid,
-//                                  to: drawingGesture.dragLocation)
-//                        .stroke(.red,
-//                                style: StrokeStyle(
-//                                    // scale DOWN when we're zoomed out, i.e. simply apply the graph scale
-//                                    lineWidth: LINE_EDGE_WIDTH * scale, //* self.document.graphMovement.zoomData,
-//                                    lineCap: .round,
-//                                    lineJoin: .round))
-                        
-                        findEligibleInput(
-                            graph.edgeDrawingObserver,
-                            draggedOutputRect: draggedOutputRect,
-                            geometry: geometry,
-                            preferences: preferences
-                        )
-                    } // if let draggedOutputPref
-                } // GeometryReader
-            } // overlayPreferenceValue
-    } // body(content:)
-
 }
 
 struct GraphHoverViewModifier: ViewModifier {
